requirements.txt
websocket-client~=1.3.2
binance~=0.3
requests